test: split palette-recolor mode and stats into their own spec
Mode routing and CPU-fallback cases no longer live in the WebGL parity file.
