+++ tests/canvas/palette-recolor-webgl_spec.js
@@ -79,106 +79,6 @@ const paletteMetadata = {
   },
 };
 
-describe("canvas/palette-recolor WebGL mode / stats / fallback", () => {
-  let previousMode;
-  let sandbox;
-
-  before(() => {
-    previousMode = getPaletteRecolorConfig().activeMode;
-  });
-
-  after(() => {
-    if (previousMode === "webgl") {
-      setPaletteRecolorMode("webgl");
-    } else {
-      setPaletteRecolorMode("cpu");
-    }
-    resetSharedWebGLForTests();
-  });
-
-  beforeEach(() => {
-    sandbox = sinon.createSandbox();
-    resetRecolorStats();
-  });
-
-  afterEach(() => {
-    if (sandbox) {
-      sandbox.restore();
-      sandbox = null;
-    }
-    resetSharedWebGLForTests();
-  });
-
-  it("increments cpu stats when forced to CPU mode", async () => {
-    setPaletteRecolorMode("cpu");
-    const img = solidCanvas(255, 0, 0);
-    const out = await recolorImage(img, [
-      { source: ["#FF0000"], target: ["#0000FF"] },
-    ]);
-
-    expect(getRecolorStats().cpu).to.be.at.least(1);
-    expect(getRecolorStats().webgl).to.equal(0);
-    expect(readPixel(out, 0, 0)).to.deep.include({ r: 0, g: 0, b: 255 });
-  });
-
-  it("resetRecolorStats clears counters", async () => {
-    setPaletteRecolorMode("cpu");
-    await recolorImage(solidCanvas(255, 0, 0), [
-      { source: ["#FF0000"], target: ["#0000FF"] },
-    ]);
-    expect(getRecolorStats().cpu).to.be.at.least(1);
-
-    resetRecolorStats();
-    expect(getRecolorStats()).to.deep.equal({
-      webgl: 0,
-      cpu: 0,
-      fallback: 0,
-    });
-  });
-
-  it("falls back to CPU with correct pixels when WebGL init fails", async function () {
-    // Need WebGL mode selected at the config level. If the browser never had
-    // WebGL, `setPaletteRecolorMode("webgl")` keeps forceCPU true — still
-    // exercise fallback by temporarily enabling useWebGL via mode after stub.
-    const config = getPaletteRecolorConfig();
-    if (!config.useWebGL && !isWebGLAvailable()) {
-      // Probe said no WebGL at module load. Still test the catch path by
-      // forcing WebGL mode attempt after stubbing context creation.
-      // `setPaletteRecolorMode("webgl")` refuses when useWebGL is false, so
-      // skip — fallback only runs when shouldUseWebGL is true.
-      this.skip();
-    }
-
-    setPaletteRecolorMode("webgl");
-    resetSharedWebGLForTests();
-
-    const originalGetContext = HTMLCanvasElement.prototype.getContext;
-    sandbox
-      .stub(HTMLCanvasElement.prototype, "getContext")
-      .callsFake(function (type, attrs) {
-        if (type === "webgl" || type === "experimental-webgl") {
-          return null;
-        }
-        return originalGetContext.call(this, type, attrs);
-      });
-
-    const warnSpy = sandbox.spy(console, "warn");
-    const img = solidCanvas(255, 0, 0);
-    const out = await recolorImage(img, [
-      { source: ["#FF0000"], target: ["#0000FF"] },
-    ]);
-
-    expect(getRecolorStats().fallback).to.be.at.least(1);
-    expect(warnSpy.called).to.equal(true);
-    expect(readPixel(out, 0, 0)).to.deep.include({
-      r: 0,
-      g: 0,
-      b: 255,
-      a: 255,
-    });
-  });
-});
-
 describe("canvas/webgl-palette-recolor.ts pixel parity", function () {
   let previousMode;
